Move Sidekiq classes into their own files
In addition, the class names have been updated to reflect their
locations
